use salesforce class in vic
