+++ lib/vic/service.rb
@@ -1,17 +1,17 @@
 # frozen_string_literal: true
 
 module VIC
-  class Service < Common::Client::Base
+  class Service < Salesforce::Service
     configuration VIC::Configuration
 
+    CONSUMER_KEY = Settings.salesforce.consumer_key
+    SIGNING_KEY_PATH = Settings.salesforce.signing_key_path
     SALESFORCE_USERNAMES = {
       'prod' => 'vetsgov-devops@listserv.gsa.gov',
       'uat' => 'vetsgov-devops@listserv.gsa.gov.uat',
       'dev' => 'vetsgov-devops@listserv.gsa.gov.vicdev'
     }.freeze
-
     SALESFORCE_USERNAME = SALESFORCE_USERNAMES[Settings.salesforce.env]
-    SALESFORCE_HOST = "https://#{Settings.salesforce.env == 'prod' ? 'login' : 'test'}.salesforce.com"
     SERVICE_BRANCHES = {
       'F' => 'Air Force',
       'A' => 'Army',
@@ -27,37 +27,6 @@ module VIC
       veteran_address
       phone
     ].freeze
-
-    def oauth_params
-      {
-        grant_type: 'urn:ietf:params:oauth:grant-type:jwt-bearer',
-        assertion: jwt_bearer_token
-      }
-    end
-
-    def jwt_bearer_token
-      JWT.encode(claim_set, private_key, 'RS256')
-    end
-
-    def claim_set
-      {
-        iss: Settings.salesforce.consumer_key,
-        sub: SALESFORCE_USERNAME,
-        aud: SALESFORCE_HOST,
-        exp: Time.now.utc.to_i.to_s
-      }
-    end
-
-    def private_key
-      OpenSSL::PKey::RSA.new(File.read(Settings.salesforce.signing_key_path))
-    end
-
-    def get_oauth_token
-      body = request(:post, '', oauth_params).body
-      Raven.extra_context(oauth_response_body: body)
-
-      body['access_token']
-    end
 
     def convert_form(form)
       converted_form = form.deep_transform_keys { |key| key.to_s.underscore }
